Make `el-get-sources` more explicit wrt versions.
I used `:checkout` (which is actually used & recognized by El-Get) if
the recipe is associated with a Git repository and the author of that
repo provides tags for the desired releases.
I used `:version` when there wouldn't be anything meaningful to
actually run `git checkout` on, but there is a specific version
of the package regardless. El-Get does not seem to mind the presence
of keys that it does not use, like `:version`.