Use Emacs (leveraging Server-Mode) as std editor.
I know that there are ways to take advantage of the
`-a`/`--alternate-editor` flags to ensure that I never have to turn on
Emacs in daemon mode manually, but I'm going to skip past getting that
set up for now. MAINLY because I'm still undecided about how to resolve
getting Smotitah to startup without any input from me.