Allow Emacs to rearrange custom.el a bit
I had been getting some cryptic error upon saving any customization
changes, so I decided to recreate the custom file from scratch via
the easy-customization GUI interface.

To elaborate a bit: the cryptic error seemed to relate to Emacs
wanting to change the indentation of the custom file (for some reason)
but then it would come across a syntax error and refuse to save the changes.